requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.13
pyjwt==2.10.1
pydantic-settings==2.10.1
pytest==8.4.1
httpx==0.28.1
